Move varargs parsers to separate namespace
