+++ src/Environments/Windows/win64characteristics.xml
@@ -2,7 +2,7 @@
 <characteristics xmlns="http://schemata.jklnet.org/Decompiler">
   <entry name="printf">
     <characteristics>
-      <varargs>Reko.Environments.Windows.MsPrintfFormatParser,Reko.Environments.Windows</varargs>
+      <varargs>Reko.Core.Analysis.PrintfFormatParser,Reko.Core</varargs>
     </characteristics>
   </entry>
   <entry name="exit">
