+++ src/UnitTests/UnitTests.csproj
@@ -551,6 +551,7 @@
     <Compile Include="Typing\UnionIntRealMock.cs" />
     <Compile Include="Typing\UnionPointersStructuresMatcherTests.cs" />
     <Compile Include="Fragments\VectorFragment.cs" />
+    <Compile Include="Core\Analysis\PrintfFormatParserTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
